requirements.base.txt
# Core
python-dotenv
typing-extensions

# API
fastapi
uvicorn
httpx
orjson
python-multipart

# UI
gradio

# Networking
requests
beautifulsoup4
feedparser

# Utilities
pydub
psutil
protobuf==3.20.3
grpcio
jsonschema
jinja2
rich
click
joblib
filelock
regex
packaging
pyyaml
